requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
defusedxml==0.7.1
dj-database-url==2.2.0
Django==3.2.25
django-allauth==0.50.0
django-braces==1.14.0
django-cors-headers==4.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
importlib-metadata==6.7.0
oauthlib==3.2.2
packaging==24.0
Pillow==8.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
whitenoise==6.5.0
zipp==3.15.0
